backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
argcomplete==1.10.3
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
autopep8==2.3.2
azure-common==1.1.28
azure-core==1.35.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
bcrypt==4.3.0
beautifulsoup4==4.8.2
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
fastapi-utils==0.8.0
fonttools==4.59.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
isodate==0.7.2
jiter==0.10.0
kiwisolver==1.4.9
lxml==6.0.1
makefun==1.16.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.3
openai==1.105.0
openpyxl==3.1.5
packaging==25.0
passlib==1.7.4
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==0.6.23
PyYAML==6.0.2
regex==2025.9.18
reportlab==4.4.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SpeechRecognition==3.8.1
SQLAlchemy==2.0.43
starlette==0.47.3
pdfminer.six==20231228
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
pytesseract==0.3.13
Pillow==11.3.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
xlrd==1.2.0
xlsxwriter==3.2.5
